+++ lib/influx.dart
@@ -1,3 +1,3 @@
-export 'src/view.dart';
-export 'src/view_manager.dart';
-export 'src/view_navigation_observer.dart';
+export 'src/data_view.dart';
+export 'src/data_view_manager.dart';
+export 'src/data_view_navigation_observer.dart';