+++ lib/src/data_view.dart
@@ -1,6 +1,6 @@
 // -- TYPES
 
-abstract class View
+abstract class DataView
 {
     // -- OPERATIONS